+++ ent/auditevent/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.AuditEvent {
+func ID(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.AuditEvent {
+func IDEQ(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.AuditEvent {
+func IDNEQ(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.AuditEvent {
+func IDIn(ids ...uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.AuditEvent {
+func IDNotIn(ids ...uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.AuditEvent {
+func IDGT(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.AuditEvent {
+func IDGTE(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.AuditEvent {
+func IDLT(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.AuditEvent {
+func IDLTE(id uint64) predicate.AuditEvent {
 	return predicate.AuditEvent(sql.FieldLTE(FieldID, id))
 }
