+++ ent/chatturn/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.ChatTurn {
+func ID(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.ChatTurn {
+func IDEQ(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.ChatTurn {
+func IDNEQ(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.ChatTurn {
+func IDIn(ids ...uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.ChatTurn {
+func IDNotIn(ids ...uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.ChatTurn {
+func IDGT(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.ChatTurn {
+func IDGTE(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.ChatTurn {
+func IDLT(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.ChatTurn {
+func IDLTE(id uint64) predicate.ChatTurn {
 	return predicate.ChatTurn(sql.FieldLTE(FieldID, id))
 }
