+++ ent/checkpoint.go
@@ -16,7 +16,7 @@ import (
 type Checkpoint struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	ID uint64 `json:"id,omitempty"`
 	// CheckpointID holds the value of the "checkpoint_id" field.
 	CheckpointID string `json:"checkpoint_id,omitempty"`
 	// Payload holds the value of the "payload" field.
@@ -63,7 +63,7 @@ func (_m *Checkpoint) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			_m.ID = int(value.Int64)
+			_m.ID = uint64(value.Int64)
 		case checkpoint.FieldCheckpointID:
 			if value, ok := values[i].(*sql.NullString); !ok {
 				return fmt.Errorf("unexpected type %T for field checkpoint_id", values[i])
