+++ ent/checkpoint/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Checkpoint {
+func ID(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Checkpoint {
+func IDEQ(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Checkpoint {
+func IDNEQ(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Checkpoint {
+func IDIn(ids ...uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Checkpoint {
+func IDNotIn(ids ...uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Checkpoint {
+func IDGT(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Checkpoint {
+func IDGTE(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Checkpoint {
+func IDLT(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Checkpoint {
+func IDLTE(id uint64) predicate.Checkpoint {
 	return predicate.Checkpoint(sql.FieldLTE(FieldID, id))
 }
