+++ ent/knowledgebase/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.KnowledgeBase {
+func ID(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.KnowledgeBase {
+func IDEQ(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.KnowledgeBase {
+func IDNEQ(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.KnowledgeBase {
+func IDIn(ids ...uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.KnowledgeBase {
+func IDNotIn(ids ...uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.KnowledgeBase {
+func IDGT(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.KnowledgeBase {
+func IDGTE(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.KnowledgeBase {
+func IDLT(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.KnowledgeBase {
+func IDLTE(id uint64) predicate.KnowledgeBase {
 	return predicate.KnowledgeBase(sql.FieldLTE(FieldID, id))
 }
